Refactor match player tracking
