+++ src/api/versions/v1/services/websocket-service.ts
@@ -9,6 +9,7 @@ import { SessionKV } from "../interfaces/kv/session-kv.ts";
 import { WebSocketType } from "../enums/websocket-enum.ts";
 import { inject, injectable } from "@needle-di/core";
 import { KVService } from "../../../../core/services/kv-service.ts";
+import { MatchPlayersService } from "./match-players-service.ts";
 import { WSMessageReceive } from "hono/ws";
 import { WebSocketUser } from "../models/websocket-user.ts";
 import { BinaryReader } from "../../../../core/utils/binary-reader-utils.ts";
@@ -21,11 +22,12 @@ export class WebSocketService {
   private serverId: string;
   private serversUserCount: Map<string, { count: number; timestamp: number }>;
   private users: Map<string, WebSocketUser>;
-  private matchPlayers: Map<string, Set<string>>;
 
-  constructor(private kvService = inject(KVService)) {
+  constructor(
+    private kvService = inject(KVService),
+    private matchPlayersService = inject(MatchPlayersService),
+  ) {
     this.users = new Map();
-    this.matchPlayers = new Map();
     this.serverId = crypto.randomUUID();
     this.broadcastChannel = new BroadcastChannel(TUNNEL_CHANNEL);
     this.onlineUsersChannel = new BroadcastChannel(ONLINE_USERS_CHANNEL);
@@ -139,7 +141,7 @@ export class WebSocketService {
     if (result.ok) {
       console.log(`Deleted temporary data for user ${userName}`);
       this.users.delete(userToken);
-      this.matchPlayers.delete(userToken);
+      this.matchPlayersService.deleteByToken(userToken);
       this.updateAndBroadcastOnlineUsers();
     } else {
       console.error(`Failed to delete temporary data for user ${userName}`);
@@ -170,7 +172,7 @@ export class WebSocketService {
       }
 
       case WebSocketType.MatchPlayer: {
-        this.handleMatchPlayerMessage(user, binaryReader);
+        this.matchPlayersService.handleMatchPlayerMessage(user, binaryReader);
         break;
       }
 
@@ -275,28 +277,6 @@ export class WebSocketService {
     this.sendMessageToOtherUser(destinationToken, playerIdentityPayload);
   }
 
-  private handleMatchPlayerMessage(
-    originUser: WebSocketUser,
-    binaryReader: BinaryReader,
-  ): void {
-    const isConnected = binaryReader.boolean();
-    const playerId = binaryReader.fixedLengthString(32);
-
-    const token = originUser.getToken();
-    let players = this.matchPlayers.get(token);
-    if (players === undefined) {
-      players = new Set<string>();
-      this.matchPlayers.set(token, players);
-    }
-
-    if (isConnected) {
-      players.add(playerId);
-      console.log(`Player ${playerId} joined match ${token}`);
-    } else {
-      players.delete(playerId);
-      console.log(`Player ${playerId} left match ${token}`);
-    }
-  }
 
   private handleTunnelMessage(
     originUser: WebSocketUser,
